refactor: Split up parser test
